Pull out vg validate
